% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_functions.R
\name{access_dwt2D}
\alias{access_dwt2D}
\title{Access a particular region from a DWT calculated with image_dwt_mult().}
\usage{
access_dwt2D(dwt2D_object, my_level, my_region)
}
\arguments{
\item{dwt2D_object}{an object returned from \code{\link[=image_dwt_mult]{image_dwt_mult()}}.}

\item{my_level}{integer. the desired level of the DWT return, where 1 is the largest level (most detailed). Note that this is the opposite of how the levels are numbered in WaveThresh (and starts at 1, not 0).}

\item{my_region}{character, to match one of c("horizontal", "vertical", "diagonal", "scaling"). will match the whole string or just initial character(s).}
}
\value{
a matrix containing the DWT coefficients from the selected level.
}
\description{
Access a particular region from a DWT calculated with image_dwt_mult().
}
