% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boat}
\alias{boat}
\title{Boat image}
\format{
\subsection{\code{boat}}{

A 512 x 512 matrix representing a grayscale image of a boat
}
}
\source{
\url{https://sipi.usc.edu/database/}
}
\usage{
boat
}
\description{
Boat image
}
\keyword{datasets}
