% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt_functions.R
\name{decomp_at_level}
\alias{decomp_at_level}
\title{perform and return a (potentially) limited number of levels of DWT decomposition with \code{\link[wavethresh:wd]{wavethresh::wd()}}, rather than the full decomposition which \code{wd()} returns be default.}
\usage{
decomp_at_level(fulldecomp, num_decomp = NULL)
}
\arguments{
\item{fulldecomp}{a \code{\link[wavethresh:wd]{wavethresh::wd()}} object}

\item{num_decomp}{integer, starting from 1. the number of levels of decomposition to return}
}
\value{
a vector containing the smooth and detail coefficients at the desired level (in that order, increasing in levels of detail).
}
\description{
perform and return a (potentially) limited number of levels of DWT decomposition with \code{\link[wavethresh:wd]{wavethresh::wd()}}, rather than the full decomposition which \code{wd()} returns be default.
}
