% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_spectra_functions.R
\name{est_H}
\alias{est_H}
\title{Estimate Hurst exponent for a 2D signal}
\usage{
est_H(slope_obj)
}
\arguments{
\item{slope_obj}{list, the output from \code{\link[=get_slope]{get_slope()}}}
}
\value{
scalar, the estimated Hurst exponent
}
\description{
Estimate Hurst exponent for a 2D signal
}
