% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbm2d_512_75}
\alias{fbm2d_512_75}
\title{2D fractional Brownian motion}
\format{
\subsection{\code{fbm2d_512_75}}{

A 512 x 512 matrix containing 2-dimensional fractional Brownian motion with a Hurst exponent of 0.75.
}
}
\source{
Generated in MATLAB using code provided by Dr. Brani Vidakovic from \url{https://project.inria.fr/fraclab/}.
}
\usage{
fbm2d_512_75
}
\description{
2D fractional Brownian motion
}
\keyword{datasets}
