% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_spectra_functions.R
\name{get_energies}
\alias{get_energies}
\title{Calculate log energies for selected levels}
\usage{
get_energies(
  dwt2D_object,
  level_vec = NULL,
  region_vec = "diagonal",
  location_stat
)
}
\arguments{
\item{dwt2D_object}{an object returned from \code{\link[=image_dwt_mult]{image_dwt_mult()}}.}

\item{level_vec}{a vector of the levels of decomposition to use}

\item{region_vec}{a single string or vector of the strings of the regious to use}

\item{location_stat}{which statistic to use for estimating the energy. currently only "mean" is supported.}
}
\value{
a list with:
\itemize{
\item \code{spectra_energies},
\item \code{level_vec},
\item \code{region_vec},
\item \code{location_stat}
}
}
\description{
Calculate log energies for selected levels
}
