% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_spectra_functions.R
\name{get_slope}
\alias{get_slope}
\title{calculate the spectra slope}
\usage{
get_slope(
  energies_obj,
  levels_select = NULL,
  level_by_index = FALSE,
  fit_method = "lm"
)
}
\arguments{
\item{energies_obj}{a list as provided by \code{\link[=get_energies]{get_energies()}}}

\item{levels_select}{a vector with a selection of the levels to use for fitting. either a list of the specific levels, or a list of the level indexes within the \code{energies_obj$level_vec} vector.}

\item{level_by_index}{whether to select the levels themselves or by index}

\item{fit_method}{currently, only ordinary least squares via \code{"lm"} is supported.}
}
\value{
a list with:
\itemize{
\item \code{fit_slope} slope of the fit line
\item \code{fit_int} intercept of the fit line
}
}
\description{
calculate the spectra slope
}
