% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HurstDCWT-package.R
\docType{package}
\name{HurstDCWT-package}
\alias{HurstDCWT}
\alias{HurstDCWT-package}
\title{HurstDCWT: Estimate Hurst's Exponent Using Discrete Complex Wavelet Transforms}
\description{
This package can be used to calculate the Hurst exponent for 2D signals such as images using discrete complex wavelet transforms.
}
\author{
\strong{Maintainer}: Ray Hinton \email{ray.hinton@gmail.com} (\href{https://orcid.org/0009-0000-1733-1613}{ORCID})

}
\keyword{internal}
