% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt_functions.R
\name{image_dwt}
\alias{image_dwt}
\title{perform DWT on a matrix, i.e. a 2D signal.}
\usage{
image_dwt(image_mat, num_decomp = NULL)
}
\arguments{
\item{image_mat}{a square matrix}

\item{num_decomp}{integer, starting from 1. the number of levels of decomposition to return}
}
\value{
a square matrix
}
\description{
perform DWT on a matrix, i.e. a 2D signal.
}
