% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwt_functions.R
\name{image_dwt_mult}
\alias{image_dwt_mult}
\title{perform multiple levels of DWT decomposition on a matrix, i.e. 2D signal.}
\usage{
image_dwt_mult(image_mat, num_decomp, ...)
}
\arguments{
\item{image_mat}{a square matrix}

\item{num_decomp}{integer, starting from 1. the number of levels of decomposition to return}

\item{...}{wavelet filter options passed on to \code{wavethresh::wd()}}
}
\value{
a list with the following
\itemize{
\item \code{dwt}
\item \code{nlevels}
}
}
\description{
perform multiple levels of DWT decomposition on a matrix, i.e. 2D signal.
}
