% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_spectra_functions.R
\name{log_e_region}
\alias{log_e_region}
\title{log energy of one region}
\usage{
log_e_region(region_mat, location_stat = "mean")
}
\arguments{
\item{region_mat}{matrix for a single region of a 2D DWCT}

\item{location_stat}{which statistic to use for estimating the energy. currently only "mean" is supported.}
}
\value{
scalar, the log2 of the estimated coefficient energy in the region
}
\description{
log energy of one region
}
