% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot the wavelet spectra}
\usage{
plot_spectra(energies_obj, slope_obj = NULL)
}
\arguments{
\item{energies_obj}{a list as provided by \code{\link[=get_energies]{get_energies()}}}

\item{slope_obj}{list, the output from \code{\link[=get_slope]{get_slope()}}}
}
\value{
a ggplot plot object
}
\description{
Plot the wavelet spectra
}
