% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_functions.R
\name{thresh_region}
\alias{thresh_region}
\title{a function to return the region label numbers used by WaveThresh}
\usage{
thresh_region(input_label)
}
\arguments{
\item{input_label}{string that specifies one of the regions in the 2D DWT. will match on the first character.}
}
\value{
an integer that corresponds to the region as labeled by WaveThresh
}
\description{
a function to return the region label numbers used by WaveThresh
}
